refactor school-related models and update UI components for consistency
