+++ app/src/main/java/com/nyansapoai/teaching/data/local/LocalDataSource.kt
@@ -1,6 +1,6 @@
 package com.nyansapoai.teaching.data.local
 
-import com.nyansapoai.teaching.data.remote.school.LocalSchoolInfo
+import com.nyansapoai.teaching.domain.models.school.LocalSchoolInfo
 import com.nyansapoai.teaching.domain.models.assessments.CompletedAssessment
 import com.nyansapoai.teaching.domain.models.assessments.literacy.PendingMultipleChoicesResult
 import com.nyansapoai.teaching.domain.models.assessments.literacy.PendingReadingAssessmentResult
