+++ app/src/main/java/com/nyansapoai/teaching/data/remote/school/SchoolRepository.kt
@@ -1,0 +1,5 @@
+package com.nyansapoai.teaching.data.remote.school
+
+interface SchoolRepository {
+    suspend fun getSchoolInfo()
+}
