+++ app/src/main/java/com/nyansapoai/teaching/domain/mapper/school/LocalSchoolInfoMapper.kt
@@ -1,6 +1,6 @@
 package com.nyansapoai.teaching.domain.mapper.school
 
-import com.nyansapoai.teaching.data.remote.school.LocalSchoolInfo
+import com.nyansapoai.teaching.domain.models.school.LocalSchoolInfo
 import database.SchoolInfoEntity
 
 fun SchoolInfoEntity.toLocalSchoolInfo(): LocalSchoolInfo {
